pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinicaltriageenv"
version = "0.1.0"
description = "AI Medical Triage Simulation Environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "requests",
    "python-dotenv",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

